Accept browsers not in requirements list
Previously, if user's browser was not in the requirements config, the bad browser page would be shown. This behaviour is undesirable because the page may be viewed by various mobile or specialized browsers, and it's hard to enumerate minimum versions for all of them.

Instead, we provide a rejection list and accept all other browsers.